+++ src/features/mypage/model/surveyAtom.jsx
@@ -1,11 +1,5 @@
 import { atom } from "recoil";
 
-// 성향조사 섹션 상태
-export const surveySectionState = atom({
-  key: 'surveySectionState',
-  default: 'proficiency',
-});
-
 // 성향조사 데이터 상태
 export const surveyState = atom({
   key: 'surveyState',
@@ -19,10 +13,4 @@ export const surveyState = atom({
     horrorPos: null,
     preference: null,
   },
-});
-
-// 모달 상태
-export const modalState = atom({
-  key: 'modalState',
-  default: false,
 });
