+++ src/features/mypage/ui/preferences/HeadcountSection.jsx
@@ -24,18 +24,18 @@ function HeadcountSection() {
         선호 인원
       </Title>
 
-        {/* 선택 영역 */}
-        <ListWrapper>
-          {[1, 2, 3, 4, 5, 6].map((num) => (
-            <List key={num}>
-              <RadioButton
-                selected={selected === num}
-                onClick={() => handleSelect(num)}
-              />
-              {num === 6 ? '6인 이상' : `${num}인`}
-            </List>
-          ))}
-        </ListWrapper>
+      {/* 선택 영역 */}
+      <ListWrapper>
+        {[1, 2, 3, 4, 5, 6].map((num) => (
+          <List key={num}>
+            <RadioButton
+              selected={selected === num}
+              onClick={() => handleSelect(num)}
+            />
+            {num === 6 ? '6인 이상' : `${num}인`}
+          </List>
+        ))}
+      </ListWrapper>
     </SectionWrapper>
   )
 }
@@ -62,36 +62,28 @@ const ListWrapper = styled.div`
   display: flex;
   flex-direction: column;
   align-items: flex-start;
-  gap: 1.25em;
+  gap: 1.25rem;
   align-self: stretch;
-
-  @media(max-width: 768px) {
-    gap: 1em;
-  }
 `;
 
 const List = styled.div`
   display: flex;
   align-items: center;
-  gap: 0.625em;
+  gap: 0.625rem;
   align-self: stretch;
   color: var(--RIU_Monochrome-300, #696C7E);
   font-family: 'Pretendard-Medium';
-  font-size: 0.875em;
+  font-size: 0.875rem;
   line-height: 130%;
-
-  @media(max-width: 768px) {
-    font-size: 0.65em;
-    line-height: 100%;
-  }
 `;
 
 const RadioButton = styled.button`
   all: unset;
   border: 1.5px solid var(--RIU_Primary-100, #718FF2);
-  border-radius: 1.875em;
-  width: 1em;
-  height: 1em;
+  border-radius: 1.875rem;
+  margin: 0.125rem;
+  width: 1rem;
+  height: 1rem;
   position: relative;
   cursor: pointer;
 
@@ -100,16 +92,12 @@ const RadioButton = styled.button`
     position: absolute;
     top: 50%;
     left: 50%;
-    width: 0.5em;
-    height: 0.5em;
+    width: 0.5rem;
+    height: 0.5rem;
     background-color: var(--RIU_Primary-100, #718FF2);
-    border-radius: 1.875em;
+    border-radius: 1.875rem;
     transform: translate(-50%, -50%);
     opacity: ${props => (props.selected ? 1 : 0)};
     transition: opacity 0.2s ease-in-out;
-  }
-
-  @media(max-width: 768px) {
-    border-width: 1px;
   }
 `;
