+++ src/features/mypage/ui/preferences/PreferenceSection.jsx
@@ -47,65 +47,65 @@ function PreferenceSection() {
 
   return (
     <SectionWrapper>
-        {/* 타이틀 */}
-        <Title>
-          숙련도
-        </Title>
+      {/* 타이틀 */}
+      <Title>
+        숙련도
+      </Title>
 
-        {/* 선택 영역 */}
-        <CheckWrapper>
-          <ListWrapper>
-            <ListTitleWrapper>
-              <ListTitle>
-                중요하게 보는 요소
-              </ListTitle>
-              <ListDescription>
-                최대 3개까지 선택 가능합니다
-              </ListDescription>
-            </ListTitleWrapper>
-            <ElementList>
-              {preferredElementList.map((item) => (
-                <SurveyTag
-                  key={item.id}
-                  item={item.value}
-                  selected={selectedElements.includes(item.enum)}
-                  onClick={() => handleElementClick(item.enum)}
-                  disabled={!selectedElements.includes(item.enum) && selectedElements.length >= 3}
-                />
-              ))}
-            </ElementList>
-          </ListWrapper>
-          <ListWrapper>
+      {/* 선택 영역 */}
+      <CheckWrapper>
+        <ListWrapper>
+          <ListTitleWrapper>
             <ListTitle>
-              장치 vs 자물쇠
+              중요하게 보는 요소
             </ListTitle>
-            <List>
-              {preferredDeviceList.map((item) => (
-                <SurveyTag
-                  key={item.id}
-                  item={item.value}
-                  selected={selectedDevice === item.enum}
-                  onClick={() => handleDeviceClick(item.enum)}
-                />
-              ))}
-            </List>
-          </ListWrapper>
-          <ListWrapper>
-            <ListTitle>
-              활동성
-            </ListTitle>
-            <List>
-              {preferredActivityList.map((item) => (
-                <SurveyTag
-                  key={item.id}
-                  item={item.value}
-                  selected={selectedActivity === item.enum}
-                  onClick={() => handleActivityClick(item.enum)}
-                />
-              ))}
-            </List>
-          </ListWrapper>
-        </CheckWrapper>
+            <ListDescription>
+              최대 3개까지 선택 가능합니다
+            </ListDescription>
+          </ListTitleWrapper>
+          <ElementList>
+            {preferredElementList.map((item) => (
+              <SurveyTag
+                key={item.id}
+                item={item.value}
+                selected={selectedElements.includes(item.enum)}
+                onClick={() => handleElementClick(item.enum)}
+                disabled={!selectedElements.includes(item.enum) && selectedElements.length >= 3}
+              />
+            ))}
+          </ElementList>
+        </ListWrapper>
+        <ListWrapper>
+          <ListTitle>
+            장치 vs 자물쇠
+          </ListTitle>
+          <List>
+            {preferredDeviceList.map((item) => (
+              <SurveyTag
+                key={item.id}
+                item={item.value}
+                selected={selectedDevice === item.enum}
+                onClick={() => handleDeviceClick(item.enum)}
+              />
+            ))}
+          </List>
+        </ListWrapper>
+        <ListWrapper>
+          <ListTitle>
+            활동성
+          </ListTitle>
+          <List>
+            {preferredActivityList.map((item) => (
+              <SurveyTag
+                key={item.id}
+                item={item.value}
+                selected={selectedActivity === item.enum}
+                onClick={() => handleActivityClick(item.enum)}
+              />
+            ))}
+          </List>
+        </ListWrapper>
+      </CheckWrapper>
     </SectionWrapper>
   )
 }
@@ -132,7 +132,7 @@ const CheckWrapper = styled.div`
   display: flex;
   flex-direction: column;
   align-items: flex-start;
-  gap: 1em;
+  gap: 1rem;
   align-self: stretch;
 `;
 
@@ -141,7 +141,7 @@ const ListWrapper = styled.div`
   flex-direction: column;
   justify-content: center;
   align-items: flex-start;
-  gap: 0.625em;
+  gap: 0.625rem;
   align-self: stretch;
 `;
 
@@ -155,47 +155,28 @@ const ListTitleWrapper = styled.div`
 const ListTitle = styled.div`
   color: var(--RIU_Monochrome-500, #515467);
   font-family: 'Pretendard-Bold';
-  font-size: 0.875em;
-  font-style: normK;
-  line-height: 1.25em;
-  letter-spacing: -0.02188em;
-
-  @media (max-width: 768px) {
-    font-size: 0.7143em;
-  }
+  font-size: 0.875rem;
+  line-height: 130%;
+  letter-spacing: -0.02188rem;
 `;
 
 const ListDescription =styled.div`
   color: var(--RIU_Monochrome-70, #B3B6C3);
   font-family: 'Pretendard-Medium';
-  font-size: 0.75em;
+  font-size: 0.75rem;
   line-height: 130%;
-
-  @media (max-width: 768px) {
-    font-size: 0.7143em;
-  }
 `;
 
 const ElementList = styled.div`
-  width: 100%;
   display: flex;
-  flex-flow: wrap;
+  width: 27.5rem;
   justify-content: space-between;
   align-items: center;
-
-  @media (max-width: 768px) {
-    width: 100%;
-    gap: 0.8125em 0.625em;
-  }
 `;
 
 const List = styled.div`
+  width: 27.5rem;
   display: flex;
-  width: 27.5em;
   align-items: center;
-  gap: 0.625em;
-
-  @media (max-width: 768px) {
-    gap: 0.625em;
-  }
+  gap: 0.625rem;
 `;
